fix types/build
