+++ packages/main/api-extractor.json
@@ -1,6 +1,9 @@
 {
   "extends": "../../config/api-extractor.json",
   "mainEntryPointFilePath": "<projectFolder>/dist/src/index.d.ts",
+  "apiReport": {
+    "reportFileName": "<unscopedPackageName>.api.md"
+  },
   "dtsRollup": {
     "enabled": true,
     "untrimmedFilePath": "<projectFolder>/dist/<unscopedPackageName>.d.ts"
