+++ packages/main/api-extractor.server.json
@@ -1,6 +1,9 @@
 {
   "extends": "../../config/api-extractor.json",
   "mainEntryPointFilePath": "<projectFolder>/dist/src/server/index.d.ts",
+  "apiReport": {
+    "reportFileName": "<unscopedPackageName>-server.api.md"
+  },
   "dtsRollup": {
     "enabled": true,
     "untrimmedFilePath": "<projectFolder>/dist/server/server.d.ts"
