+++ packages/main/types/server/index.ts
@@ -18,3 +18,7 @@
 export * from './files';
 export * from './caching';
 export * from './shared';
+
+export { RequestOptions } from "../../types/requests";
+export * from "../../types/content";
+export * from "../../types/cached-content";
